+++ lib/pipeline/types.ex
@@ -4,32 +4,24 @@ defmodule Pipeline.Types do
   """
 
   @typedoc """
+  A reference to a step/hook.
+
+  It's a two-element tuple, where the first element is the module name and the second element is the function name.
+  """
+  @type step_ref :: {module(), atom()}
+
+  @typedoc """
   The result of the execution of a step or the final result of an entire pipeline.
   """
-  @type result :: {:ok, any} | {:error, any}
+  @type result :: {:ok, any()} | {:error, any()}
 
   @typedoc """
   The arguments that are used by the first step of a pipeline and well - the initial state of the pipeline.
   """
-  @type args :: any
+  @type args :: any()
 
   @typedoc """
   The optional arguments that are passed to steps and callbacks of a pipeline.
   """
   @type options :: Keyword.t()
-
-  @typedoc """
-  A function that transforms a value into something else.
-  """
-  @type reducer :: (args(), options() -> result())
-
-  @typedoc """
-  A function that is always called after the pipeline finishes the execution.
-  """
-  @type hook :: (Pipeline.State.t(), options -> any())
-
-  @typedoc """
-  A function that is always called after the pipeline finishes the execution, but asynchronously.
-  """
-  @type async_hook :: hook()
 end